[#10] refactor: types/users를 types/user로 네이밍 변경
