+++ src/types/index.ts
@@ -1,2 +1,2 @@
 export * from './auth';
-export * from './users';
+export * from './user';
